Initial commit for basic core apps functionality. Bundled core apps will be stored in an apps directory under the Cytoscape installation directory; updates to core apps will go in ~/CytoscapeConfiguration/3/apps/installed
For now, core apps continue to live in cytoscape-impl repository (but
are not included in features.xml).This implementation does not depend on
an Internet connection for initial startup, and will choose the right
version of the core app to be sued (the one in the installation
directory or apps/installed) based on what it the newest compatible
release.

This commit depends on the related commit to cytoscape-impl - without
it, the core apps will not load on startup.

Refs #3118