Changes to prevent writing to protected directories in the install
The jetty feature generates it's own config file if it isn't present, so I double checked that the file generated on my version didn't contain any system-specific settings (it didn't) and made it part of the Cytoscape distribution. We can revisit that file later if there's something we need Jetty to do, but for now, it works with the default settings.

Karaf now generates a karad.pid file whenever it starts, so I had to add a setting in custom.properties that sets the location to the generated CytoscapeConfiguration directory.

Karaf's shell scripts are somewhat more distributed now, and required a change to include several, and not the single one it used to be.